fix: move toTitleCase to string.utils and update imports
